# perf: update template renderer paths to use package qualifiers