--- a/warehouse/admin/views/prohibited_project_names.py
+++ b/warehouse/admin/views/prohibited_project_names.py
@@ -27,7 +27,7 @@ from warehouse.utils.project import prohibit_and_remove_project
 
 @view_config(
     route_name="admin.prohibited_project_names.list",
-    renderer="admin/prohibited_project_names/list.html",
+    renderer="warehouse.admin:templates/admin/prohibited_project_names/list.html",
     permission=Permissions.AdminProhibitedProjectsRead,
     request_method="GET",
     uses_session=True,
@@ -70,7 +70,7 @@ def prohibited_project_names(request):
 
 @view_config(
     route_name="admin.prohibited_project_names.add",
-    renderer="admin/prohibited_project_names/confirm.html",
+    renderer="warehouse.admin:templates/admin/prohibited_project_names/confirm.html",
     permission=Permissions.AdminProhibitedProjectsWrite,
     request_method="GET",
     uses_session=True,
@@ -285,7 +285,7 @@ def remove_prohibited_project_names(request):
 
 @view_config(
     route_name="admin.prohibited_project_names.bulk_add",
-    renderer="admin/prohibited_project_names/bulk.html",
+    renderer="warehouse.admin:templates/admin/prohibited_project_names/bulk.html",
     permission=Permissions.AdminProhibitedProjectsWrite,
     uses_session=True,
     require_methods=False,