… warehouse/manage/views/organizations.py
… warehouse/manage/views/teams.py
… warehouse/organizations/views.py
… warehouse/packaging/views.py
… warehouse/rss/views.py
… warehouse/sitemap/views.py
… warehouse/views.py