@brettlangdon 2025-06-16
What should we do about: https://github.com/pypi/warehouse/blob/3a1f8c86e59a4d54bfaf2ce6c42d59d918730600/warehouse/config.py#L728-L733

Is this proposed change even desired for the minimal gain vs having the default search path?