@brettlangdon 2025-06-16
An alternative here is to move all the templates into their respective module paths, e.g. `warehouse/templates/api/simple/detail.html` -> `warehouse/api/templates/api/simple/detail.html` and then update `renderer="templates/simple/detail.html" (since `warehouse.api:` will get prepended to all template lookups by default which don't have `:` in the name)