@miketheman 2025-07-14
https://github.com/pypi/warehouse/pull/18366 likely solves the concern, so once that lands, this PR should be checked against that for accuracy.