@miketheman 2025-07-14
@brettlangdon looks like the checker I wrote doesn't yet account for the updated paths - would you want to take a stab at adding that?